+++ tt/tests/unit/trees/_helpers.py
@@ -1,4 +1,4 @@
-"""Helpers for testing expression trees and nodes."""
+"""Helpers for testing expression nodes."""
 
 import unittest
 
@@ -9,4 +9,4 @@ class ExpressionTreeAndNodeTestCase(unittest.TestCase):
 
     def get_tree_root_from_expr_str(self, expr_str):
         """Get an tree root node from an expression string."""
-        return BooleanExpression(expr_str).tree.root
+        return BooleanExpression(expr_str).tree