+++ tt/tests/unit/trees/test_node_build_tree_exceptions.py
@@ -1,26 +1,26 @@
-"""Test tree exceptions that occur during initialization."""
+"""Test exceptions that occur in ExpressionTreeNode.build_tree."""
 
 import unittest
 
 from tt.errors import (
     InvalidArgumentTypeError,
     InvalidArgumentValueError)
-from tt.trees import BooleanExpressionTree
+from tt.trees import ExpressionTreeNode
 
 
-class TestTreeInitExceptions(unittest.TestCase):
+class TestNodeBuildTreeExceptions(unittest.TestCase):
 
     def test_postfix_tokens_not_a_list(self):
         """Test passing a non-list for postfix_tokens."""
         with self.assertRaises(InvalidArgumentTypeError):
-            BooleanExpressionTree('should cause an exception')
+            ExpressionTreeNode.build_tree('should cause an exception')
 
     def test_postfix_tokens_contains_non_str(self):
         """Test passing a list containing a non-str for postfix_tokens."""
         with self.assertRaises(InvalidArgumentTypeError):
-            BooleanExpressionTree(['A', 1, 'or'])
+            ExpressionTreeNode.build_tree(['A', 1, 'or'])
 
     def test_postfix_tokens_empty(self):
         """Test passing an empty list for postfix_tokens."""
         with self.assertRaises(InvalidArgumentValueError):
-            BooleanExpressionTree([])
+            ExpressionTreeNode.build_tree([])