+++ tt/trees/__init__.py
@@ -1,6 +1,5 @@
 """Tools for working with Boolean expression trees."""
 
-from .expr_tree import BooleanExpressionTree  # noqa
 from .tree_node import (  # noqa
     BinaryOperatorExpressionTreeNode,
     ExpressionTreeNode,