+++ docs/release_notes.rst
@@ -9,12 +9,16 @@ Check below for new features added in each release. Please note that release not
 
 Features in the 0.6.x series of releases are focused on expanding functionality to include expression optimization, satisfiability, and transformations.
 
+0.6.2 (in progress)
+```````````````````
+
+    * Remove class ``BooleanExpressionTree`` in favor of working exclusively with instances of :class:`ExpressionTreeNode <tt.trees.tree_node.ExpressionTreeNode>`
+
 0.6.1
 `````
-
     * Add iff (``iff``, ``->``) and implies (``impl``, ``<->``) Boolean operators
     * Add :data:`is_cnf <tt.expressions.bexpr.BooleanExpression.is_cnf>` and :data:`is_dnf <tt.expressions.bexpr.BooleanExpression.is_dnf>` attributes to :class:`BooleanExpression <tt.expressions.bexpr.BooleanExpression>`
-    * Add functionality to initialize :class:`BooleanExpression <tt.expressions.bexpr.BooleanExpression>` objects from instances of :class:`ExpressionTreeNode <tt.trees.tree_node.ExpressionTreeNode>` or :class:`BooleanExpressionTree <tt.trees.expr_tree.BooleanExpressionTree>`
+    * Add functionality to initialize :class:`BooleanExpression <tt.expressions.bexpr.BooleanExpression>` objects from instances of :class:`ExpressionTreeNode <tt.trees.tree_node.ExpressionTreeNode>` or ``BooleanExpressionTree``
     * Update __str__ and __repr__ for :class:`BooleanExpression <tt.expressions.bexpr.BooleanExpression>`
     * Add :data:`is_really_unary <tt.trees.tree_node.ExpressionTreeNode.is_really_unary>` attribute to :class:`ExpressionTreeNode <tt.trees.tree_node.ExpressionTreeNode>`
     * Add :func:`iter_clauses <tt.trees.tree_node.ExpressionTreeNode.iter_clauses>`, :func:`iter_cnf_clauses <tt.trees.tree_node.ExpressionTreeNode.iter_cnf_clauses>`, and :func:`iter_dnf_clauses <tt.trees.tree_node.ExpressionTreeNode.iter_dnf_clauses>` to :class:`ExpressionTreeNode <tt.trees.tree_node.ExpressionTreeNode>`